day 9
Replaced reduce with scan.

Scan allows an early termination.

scan is an iterator adapter, which means that it wraps an iterator and
calls next() on it, but it can decide to stop iteration by returning
None to its caller. (In this case the caller is count() that follows
scan()).

In this particular case, the scan's state is the "lead" point and it is
updated in each iteration with the current "follower" whenever there is
a non-zero move. If there is no move, then scan() stops by returning
None.